chore(coordinator): make persistence module flat
