+++ coordinator/app/build.gradle
@@ -37,11 +37,7 @@ dependencies {
   implementation project(':coordinator:ethereum:message-anchoring')
   implementation project(':coordinator:ethereum:forced-transactions')
   implementation project(':coordinator:clients:web3signer-client')
-  implementation project(':coordinator:persistence:blob')
-  implementation project(':coordinator:persistence:aggregation')
-  implementation project(':coordinator:persistence:batch')
-  implementation project(':coordinator:persistence:feehistory')
-  implementation project(':coordinator:persistence:db-common')
+  implementation project(':coordinator:persistence')
 
   implementation project(':coordinator:ethereum:gas-pricing:static-cap')
   implementation project(':coordinator:ethereum:gas-pricing:dynamic-cap')
