+++ coordinator/core/build.gradle
@@ -24,7 +24,7 @@ dependencies {
   testImplementation(project(':jvm-libs:generic:logging'))
   testImplementation(project(':jvm-libs:linea:metrics:micrometer'))
   testImplementation(testFixtures( project(':coordinator:core-interfaces')))
-  testImplementation(testFixtures(project(':coordinator:persistence:aggregation')))
+  testImplementation(testFixtures(project(':coordinator:persistence')))
   testImplementation(testFixtures(project(':jvm-libs:linea:clients:interfaces')))
   testImplementation(testFixtures(project(':jvm-libs:linea:core:traces')))
   testImplementation(testFixtures(project(':jvm-libs:generic:extensions:kotlin')))
