+++ coordinator/ethereum/blob-submitter/build.gradle
@@ -10,7 +10,7 @@ dependencies {
   implementation(project(":coordinator:clients:smart-contract-client"))
   implementation(project(':coordinator:ethereum:common'))
   implementation(project(':coordinator:ethereum:gas-pricing'))
-  implementation(project(":coordinator:persistence:blob"))
+  implementation(project(":coordinator:persistence"))
   implementation("com.fasterxml.jackson.core:jackson-databind:${libs.versions.jackson.get()}")
   implementation("com.fasterxml.jackson.core:jackson-annotations:${libs.versions.jackson.get()}")
   implementation("com.fasterxml.jackson.module:jackson-module-kotlin:${libs.versions.jackson.get()}")
@@ -22,8 +22,7 @@ dependencies {
 
   testImplementation(project(":jvm-libs:linea:testing:l1-blob-and-proof-submission"))
   testImplementation(project(':coordinator:ethereum:test-utils'))
-  testImplementation(project(":coordinator:persistence:aggregation"))
-  testImplementation(project(":coordinator:persistence:db-common"))
+  testImplementation(project(":coordinator:persistence"))
   testImplementation(testFixtures(project(":coordinator:ethereum:gas-pricing")))
   testImplementation(testFixtures(project(":coordinator:core-interfaces")))
   testImplementation(testFixtures(project(":jvm-libs:generic:extensions:kotlin")))
