+++ coordinator/ethereum/forced-transactions/build.gradle
@@ -4,7 +4,7 @@ plugins {
 
 dependencies {
   api project(":coordinator:core")
-  api project(":coordinator:persistence:aggregation")
+  api project(":coordinator:persistence")
   api project(":jvm-libs:linea:clients:linea-contract-clients")
   api project(":jvm-libs:linea:clients:interfaces")
   api project(":jvm-libs:linea:clients:eth-logs-searcher")
@@ -18,5 +18,5 @@ dependencies {
   testImplementation testFixtures(project(":jvm-libs:linea:core:domain-models"))
   testImplementation testFixtures(project(":jvm-libs:linea:clients:interfaces"))
   testImplementation testFixtures(project(":jvm-libs:linea:clients:linea-contract-clients"))
-  testImplementation testFixtures(project(":coordinator:persistence:aggregation"))
+  testImplementation testFixtures(project(":coordinator:persistence"))
 }
