--- coordinator/persistence/aggregation/build.gradle
@@ -1,43 +1,0 @@
-plugins {
-  id "net.consensys.zkevm.kotlin-library-conventions"
-  id "java-test-fixtures"
-}
-
-dependencies {
-  api(project(":coordinator:core-interfaces"))
-  api(project(":coordinator:clients:prover-client:serialization"))
-  implementation(project(":jvm-libs:generic:persistence:db"))
-  implementation(project(":coordinator:persistence:batch"))
-
-  testImplementation(project(":coordinator:persistence:batch"))
-  testImplementation(project(":coordinator:persistence:blob"))
-  testImplementation(project(":coordinator:persistence:db-common"))
-  testImplementation "io.consensys.tuweni:tuweni-units:${libs.versions.tuweni.get()}"
-  testImplementation(testFixtures(project(":coordinator:core-interfaces")))
-  testImplementation(testFixtures(project(":coordinator:persistence:db-common")))
-  testImplementation(testFixtures(project(":jvm-libs:generic:persistence:db")))
-  testImplementation(testFixtures(project(":jvm-libs:generic:extensions:kotlin")))
-  testImplementation("io.vertx:vertx-junit5")
-}
-
-sourceSets {
-  integrationTest {
-    kotlin {
-      compileClasspath += main.output
-      runtimeClasspath += main.output
-    }
-    compileClasspath += sourceSets.main.output + sourceSets.main.compileClasspath + sourceSets.test.compileClasspath
-    runtimeClasspath += sourceSets.main.output + sourceSets.main.runtimeClasspath + sourceSets.test.runtimeClasspath
-  }
-}
-
-task integrationTest(type: Test) { test ->
-  description = "Runs integration tests."
-  group = "verification"
-  useJUnitPlatform()
-
-  classpath = sourceSets.integrationTest.runtimeClasspath
-  testClassesDirs = sourceSets.integrationTest.output.classesDirs
-
-  dependsOn(":localStackPostgresDbOnlyComposeUp")
-}
