--- coordinator/persistence/batch/src/integrationTest/resources/valid-transaction.rlp
@@ -1,1 +1,0 @@
-0x02f9041383f1754480843b9aca00843b9aca0082520894ca896e096d359bf9b506c713d814e1088d45ed1380b903a4b7cbaede0000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000030000000000000000000000000000000000000000000000000000000000000000004f704ef38853675bdde741c3e069e80cff9fda92ff8349efeef543f23ab2bd5100000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000020a89dd44bed659cf10b72df0ce810066ff8270dd48f77f833dbfedff85353e26200000000000000000000000000000000000000000000000000000000647dfbf900000000000000000000000000000000000000000000000000000000000000a00000000000000000000000000000000000000000000000000000000000000180000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000010100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000102000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000021095000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000064351f6b6dc3efb401f1a5739629717f5a4581e3182115fbfceb15b604f4a9025418bdae02584457db3fb5aaa03568a4d4c549ea55331fe94122baa94966a8d002ea56a04b789dc2dc630cb456aae30c18a88463841743d5d6f6b4338207806dc9032f7b9300000000000000000000000000000000000000000000000000000000c001a0c768a0f861c0008b85b2aac40aaa4403f67625e7fdeb4470b169db9133927ffba04cc07a8c85e708df7afae10c679d5edb4468e7b74e85a1bb925b93008f548d41
