--- coordinator/persistence/blob/build.gradle
@@ -1,55 +1,0 @@
-import org.gradle.api.tasks.testing.logging.TestExceptionFormat
-import org.gradle.api.tasks.testing.logging.TestLogEvent
-
-plugins {
-  id "net.consensys.zkevm.kotlin-library-conventions"
-}
-
-dependencies {
-  api(project(":coordinator:core-interfaces"))
-  api(project(":coordinator:clients:prover-client:serialization"))
-  implementation(project(":jvm-libs:generic:persistence:db"))
-
-  testImplementation("com.fasterxml.jackson.core:jackson-annotations:${libs.versions.jackson.get()}")
-  testImplementation("com.fasterxml.jackson.core:jackson-databind:${libs.versions.jackson.get()}")
-  testImplementation("com.fasterxml.jackson.module:jackson-module-kotlin:${libs.versions.jackson.get()}")
-  testImplementation "io.consensys.tuweni:tuweni-units:${libs.versions.tuweni.get()}"
-  testImplementation(project(":coordinator:persistence:db-common"))
-  testImplementation(testFixtures(project(":coordinator:core-interfaces")))
-  testImplementation(testFixtures(project(":jvm-libs:generic:extensions:kotlin")))
-  testImplementation(testFixtures(project(":jvm-libs:generic:persistence:db")))
-  testImplementation("io.vertx:vertx-junit5")
-}
-
-sourceSets {
-  integrationTest {
-    kotlin {
-      compileClasspath += main.output
-      runtimeClasspath += main.output
-    }
-    compileClasspath += sourceSets.main.output + sourceSets.main.compileClasspath + sourceSets.test.compileClasspath + project(":coordinator:core-interfaces").sourceSets.main.output + project(":coordinator:core").sourceSets.main.output
-    runtimeClasspath += sourceSets.main.output + sourceSets.main.runtimeClasspath + sourceSets.test.runtimeClasspath + project(":coordinator:core-interfaces").sourceSets.main.output + project(":coordinator:core").sourceSets.main.output
-  }
-}
-
-task integrationTest(type: Test) { test ->
-  testLogging {
-    events TestLogEvent.FAILED,
-        TestLogEvent.SKIPPED,
-        TestLogEvent.STANDARD_ERROR,
-        TestLogEvent.STANDARD_OUT
-    exceptionFormat TestExceptionFormat.FULL
-    showCauses true
-    showExceptions true
-    showStackTraces true
-    showStandardStreams false
-  }
-  description = "Runs integration tests."
-  group = "verification"
-  useJUnitPlatform()
-
-  classpath = sourceSets.integrationTest.runtimeClasspath
-  testClassesDirs = sourceSets.integrationTest.output.classesDirs
-
-  dependsOn(":localStackPostgresDbOnlyComposeUp")
-}
