--- coordinator/persistence/feehistory/src/integrationTest/resources/log4j2.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="warn">
-    <Appenders>
-        <Console name="console" target="SYSTEM_OUT">
-            <PatternLayout pattern="[%-5level] %d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %c{1} - %msg%n"/>
-        </Console>
-    </Appenders>
-    <Loggers>
-        <Logger name="net.consensys.zkevm.persistence.dao.feehistory" level="trace" additivity="false">
-          <AppenderRef ref="console"/>
-        </Logger>
-        <Root level="info" additivity="false">
-            <appender-ref ref="console"/>
-        </Root>
-    </Loggers>
-</Configuration>
