+++ coordinator/persistence/src/integrationTest/resources/log4j2.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Configuration status="warn">
+    <Appenders>
+        <Console name="console" target="SYSTEM_OUT">
+            <PatternLayout pattern="[%-5level] %d{yyyy-MM-dd HH:mm:ss.SSS} [%t] %c{20} - %msg%n"/>
+        </Console>
+    </Appenders>
+    <Loggers>
+        <Logger name="net.consensys.zkevm.persistence" level="trace" additivity="false">
+            <AppenderRef ref="console"/>
+        </Logger>
+        <Logger name="net.consensys.zkevm.ethereum" level="trace" additivity="false">
+            <AppenderRef ref="console"/>
+        </Logger>
+        <Root level="warn" additivity="false">
+            <appender-ref ref="console"/>
+        </Root>
+    </Loggers>
+</Configuration>
