+++ coordinator/persistence/src/main/resources/db/V001__initial_schema.sql
