+++ coordinator/persistence/src/main/resources/db/V003__add_feehistory_schema.sql
