+++ coordinator/persistence/src/main/resources/db/V004__remove_versioning_schema.sql
