+++ coordinator/persistence/src/main/resources/db/V005__add_forced_transactions_schema.sql
