+++ settings.gradle
@@ -56,11 +56,7 @@ include 'coordinator:ethereum:forced-transactions'
 include 'coordinator:ethereum:blob-submitter'
 include 'coordinator:ethereum:message-anchoring'
 include 'coordinator:ethereum:test-utils'
-include 'coordinator:persistence:blob'
-include 'coordinator:persistence:aggregation'
-include 'coordinator:persistence:batch'
-include 'coordinator:persistence:feehistory'
-include 'coordinator:persistence:db-common'
+include 'coordinator:persistence'
 include 'coordinator:ethereum:gas-pricing'
 include 'coordinator:ethereum:gas-pricing:static-cap'
 include 'coordinator:ethereum:gas-pricing:dynamic-cap'
